requirements.txt
fastapi
uvicorn[standard]
httpx
sentence-transformers
faiss-cpu
sqlmodel
python-dotenv
